introduce lists of vnames
these lists exist for each indentation level. By appending the separator
(only the separator) the lists can be accessed. They hold a list of
child members, which can be used to iterate over these variables.
Note that if the separator is the default separator (_) one would
assign to $_ on the topl-level, which is a special variable in bash.
In this special case we assign to $__.