two bugfixes for comments and lists of values
comments were not filtered out correctly when quoted strings were around.
This may still be the case when quotes are used inside and outside of the comment.
Hoever, the case when quotes were only inside the comment should be treated correctly now.

The keys for multiline strings were added several times to the lists of values.